[Feat]: add helper text component and rename the password input and remove email regex and use yup instead
