+++ src/components/loginForm/loginForm.constants.ts
@@ -1,11 +1,11 @@
 import { PASSWORD_MAX_LENGTH, PASSWORD_MIN_LENGTH } from "@src/utils/constants";
-import { EMAIL_REGEX, PASSWORD_REGEX } from "@src/utils/regex";
+import { PASSWORD_REGEX } from "@src/utils/regex";
 import * as Yup from "yup";
 
 export const LOGIN_FORM_VALIDATION_SCHEMA = Yup.object().shape({
   email: Yup.string()
     .required(`Enter your email address`)
-    .matches(EMAIL_REGEX, `Enter a valid email address`),
+    .email(`Enter a valid email address`),
   password: Yup.string()
     .required(`Enter your password`)
     .min(
@@ -18,6 +18,6 @@ export const LOGIN_FORM_VALIDATION_SCHEMA = Yup.object().shape({
     )
     .matches(
       PASSWORD_REGEX,
-      `Your password must contain at least one character or Symbol, and a number`,
+      `Your password must contain at least one Uppercase character or Symbol, and a number`,
     ),
 });
