+++ src/components/ui/helperText/helperText.module.scss
@@ -1,0 +1,10 @@
+.helper {
+  align-self: flex-start;
+  font-size: 0.8rem;
+  font-weight: 300;
+  padding-left: 0.5rem;
+  margin-top: -0.5rem;
+}
+.error {
+  color: $error300;
+}
