+++ src/components/ui/helperText/helperText.types.ts
@@ -1,0 +1,4 @@
+export interface HelperTextProps {
+  className?: string;
+  error?: boolean;
+}
